DROID-3308 fix
